Remove default-data-readers namespace.
The default data readers contain support for #inst and #uuid reader
literals. For bootstrapping purposes this is not needed and it could be
argued that support for this is actually the responsibility of
ClojureScript, not tools.reader. The actually functionality needed for
getting ClojureScript up and running is a tools.reader light which does
not contain the default data readers.