Remove default data readers.
Remove support for default data readers as these are not needed for
bootstrapping the ClojureScript compiler. The readers should be
re-implemented in ClojureScript using JavaScript datatypes before the
ClojureScript compiler can be considered fully functional.

To me it is not obvious if the readers should be implemented in the
reader or in the ClojureScript compiler.